requirements.txt
streamlit==1.37.1
beautifulsoup4==4.12.3
requests==2.32.3
selenium==4.22.0
webdriver-manager==4.0.2
pandas==2.2.2
Pillow==10.4.0
tqdm==4.66.4
python-dotenv==1.0.1
lxml==5.2.2
